baselines/agent-based/vultrial/MultiAgents/VulTrial/requirements.txt
pyyaml
fastapi==0.95.1
uvicorn
py3langid
setuptools-scm
openai==1.1.0
opencv-python==4.8.0.76
gradio
httpx[socks]==0.25.0
astunparse
langchain==0.0.157
scikit-learn
pydantic==1.10.7
typing-extensions==4.5.0
typing-inspect==0.8.0
colorlog
rapidfuzz
spacy
colorama==0.4.6
tiktoken==0.5.1
